Asistencia_Alumnos2/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.11
django-filter==24.2
pillow==10.3.0
psycopg2==2.9.9
sqlparse==0.5.0
typing-extensions==4.11.0
tzdata==2024.1
